add some function

+++ inc/Filter.h
@@ -1,4 +1,4 @@
 #include "GrayBMP.h"
 
-void Filter_Gaussian(GrayBMP &src,GrayBMP &dst, int Range, double Var);
+void Filter_Gaussian(GrayBMP &src,GrayBMP &dst, int Range, float Var);
 void Filter_Median(GrayBMP &src,GrayBMP &dst, int Range);
